feat: refactor naming convention
