+++ src/Api/Controllers/FoldersController.cs
@@ -1,6 +1,6 @@
 using Application.Common.Models;
 using Application.Common.Models.Dtos.Physical;
-using Application.Folders.Commands.CreateFolder;
+using Application.Folders.Commands.AddFolder;
 using Microsoft.AspNetCore.Mvc;
 
 namespace Api.Controllers;
@@ -8,7 +8,7 @@ namespace Api.Controllers;
 public class FoldersController : ApiControllerBase
 {
     [HttpPost]
-    public async Task<ActionResult<Result<FolderDto>>> CreateFolders(CreateFolderCommand command)
+    public async Task<ActionResult<Result<FolderDto>>> AddFolders(AddFolderCommand command)
     {
         var result = await Mediator.Send(command);
         return Ok(Result<FolderDto>.Succeed(result));
